feat: 스낵바 swipe 로직을 CustomSnackbar로 이동
