+++ src/components/layout/RootLayout/components/A2hsSnackbar/style.ts
@@ -1,20 +1,14 @@
 import CustomButton from '@/components/CustomMui/CustomButton';
 import CustomSnackbar from '@/components/CustomMui/CustomSnackbar';
-import { MAX_WIDTH } from '@/constants/css/value';
 import styled from '@emotion/styled';
 
 const Snackbar = styled(CustomSnackbar)`
   position: fixed;
   bottom: 80px;
-  width: 100%;
-  max-width: ${MAX_WIDTH};
-  padding: 0 16px;
-  touch-action: none;
-  cursor: grab;
-  user-select: none;
 
   & .MuiPaper-root {
     display: flex;
+    flex-wrap: nowrap;
     justify-content: space-between;
     width: 100%;
     height: 42px;
